+++ app/src/main/java/com/limefriends/molde/ui/mypage/comment/MyCommentActivity.java
@@ -11,11 +11,10 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.AbsListView;
-import android.widget.ExpandableListView;
 import android.widget.ProgressBar;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.limefriends.molde.R;
 import com.limefriends.molde.comm.Constant;
@@ -23,6 +22,7 @@ import com.limefriends.molde.comm.MoldeApplication;
 import com.limefriends.molde.comm.custom.addOnListview.AddOnScrollExpandableListView;
 import com.limefriends.molde.comm.custom.addOnListview.AddOnScrollRecyclerView;
 import com.limefriends.molde.comm.custom.addOnListview.OnLoadMoreListener;
+import com.limefriends.molde.comm.utils.NetworkUtil;
 import com.limefriends.molde.comm.utils.PreferenceUtil;
 import com.limefriends.molde.entity.FromSchemaToEntitiy;
 import com.limefriends.molde.entity.comment.CommentEntity;
@@ -56,63 +56,6 @@ import static com.limefriends.molde.comm.Constant.Comment.EXTRA_KEY_COMMENT_DESC
 import static com.limefriends.molde.comm.Constant.Common.EXTRA_KEY_CARDNEWS_ID;
 import static com.limefriends.molde.comm.Constant.Common.PREF_KEY_AUTHORITY;
 
-
-/**
- * 정리
- * <p>
- * 서버에서 데이터를 묶어서 보내주는 게 가장 편하긴 하지만 클라이언트 사정에 따라 일일이 서버를 구성하면
- * <p>
- * 확장성이 떨어지고 1:1 코드가 늘어나기 때문에 서버에서는 데이터 호출을 모듈화 해 놓는 듯 하다.
- * <p>
- * 따라서 클라이언트에서 2개 api 호출이 필요한 경우, 또한 2개의 api 를 묶어서 보여줘야 하는 경우 보다 복잡한 구성이 이루어진다
- * <p>
- * 1. 첫번째 api 호출 이후 반복문을 돌면서 다시 api 를 호출하는 경우
- * <p>
- * retrofit 은 비동기 호출이기 때문에 10개 반복문을 돌면 10개 응답 중 어떤 것이 마지막에 올지 예측할 수 없다
- * <p>
- * 따라서 해당 호출에 count++ 를 사용한다거나 size() 를 통해 비교한다거나 멤버변수 값을 변화시키려고 하는 경우
- * <p>
- * 혹은 일부는 반복문 내부에서 돌리고 나머지는 네트워크의 응답을 통해 값을 변경하려고 하는 경우
- * <p>
- * 눈에는 속도가 비슷해 보여도 네트워크 속도가 훨씬 느리기 때문에 반복문이 전부 먼저 실행되고 나머지 네트워크가 한참 뒤에
- * <p>
- * 그것도 어떤 것이 먼저인지 모른 채로 응답된다.
- * <p>
- * 참고로 첫번째 케이스는 카드뉴스와 댓글을 하나의 List 에 담아 넘겨주는 형식으로 함
- * <p>
- * 2. 첫번째 api 호출 이후 반복문을 돌면서 호출하되 값을 변수로 넘겨줘서 콜백에서 값을 묶어주는 경우
- * <p>
- * 이렇게 할 경우 카드뉴스 객체에 댓글 리스트를 담아 ExpandableListView 에서 보여주는 형태가 된다
- * <p>
- * 3. 첫번째 api 호출 이후 반복문을 돌면서 호출하되 count 값에 락을 걸어서 마지막에 값을 갱신하도록 하는 경우
- * <p>
- * 이렇게 할 경우 count 값에 반드시 락을 걸어서 비교하려고 값을 가져오고 값을 연산하는 사이에 값이 또 변경되는 경우가 없도록 해야 한다
- * <p>
- * 4. 동기로 데이터를 불러오는 경우
- * <p>
- * enqueue 가 아닌 execute 혹은 따로 스레드를 통해서 하나씩 값을 불러올 수 있다.
- * <p>
- * 혹은 synchronized 를 통해 하나씩 데어터를 호출하도록 할 수 있다.
- * <p>
- * 결론)
- * <p>
- * 어떤 경우에 10개 이상의 비동기 처리를 생각해 줘야 하는지
- * <p>
- * 비동기로 호출한 2개의 값을 어떻게 하나로 합칠 수 있는지 생각해 보면
- * <p>
- * 왜 synchronized 가 필요한지 Observable, 스레드 처리, Retrofit, OkHttp, Rx 가
- * <p>
- * 왜 사용되는지 왜 등장했는지 조금은 이해할 수 있을 것 같다.
- * <p>
- * 다양한 케이스를 생각해 보면 결국 어디에서는 특정 시점을 캐치해 낼 수 있어야 하고
- * <p>
- * 어디에선가는 동기화가 필요하다.
- * <p>
- * 마지막으로 Rx 를 사용하면 정말 편하게 개발할 수 있을 것 같다는 생각이 듬
- */
-// TODO responsecount로 처리하는 애들 중에 만약 중간에 네트워크가 끊기거나 하나 못 받아오면 어떻게 되는거임
-// TODO addOnScroll 추가 구현할 것
-// TODO 어드민 페이지랑 delete api 추가할 것
 public class MyCommentActivity extends AppCompatActivity implements MyCommentExpandableAdapter.OnItemClickCallback {
 
     @BindView(R.id.myComment_container)
@@ -191,7 +134,6 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
             myComment_reported_comment_listview.setOnLoadMoreListener(new OnLoadMoreListener() {
                 @Override
                 public void loadMore() {
-                    Log.e("호출확인 ", "loadMore3");
                     loadReportedComment(PER_PAGE, currentPage);
                 }
             });
@@ -255,6 +197,11 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
 
     // 1. 내가 작성한 댓글
     private void loadComment(int perPage, int page) {
+
+        if (!NetworkUtil.isConnected(this)) {
+            Toast.makeText(this, "인터넷 연결을 확인해주세요.", Toast.LENGTH_LONG).show();
+            return;
+        }
 
         if (!hasMoreToLoad) return;
 
@@ -418,6 +365,11 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
     // 2. 관리자용 신고된 댓글
     private void loadReportedComment(int perPage, int page) {
 
+        if (!NetworkUtil.isConnected(this)) {
+            Toast.makeText(this, "인터넷 연결을 확인해주세요.", Toast.LENGTH_LONG).show();
+            return;
+        }
+
         if (!hasMoreToLoad) return;
 
         myComment_reported_comment_listview.setIsLoading(true);
@@ -506,18 +458,20 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
     // 3. 댓글 삭제
     public void deleteComment(final int position, final int commentId) {
 
-        String uId
-                = ((MoldeApplication) getApplication()).getFireBaseAuth().getCurrentUser().getUid();
+        if (!NetworkUtil.isConnected(this)) {
+            Toast.makeText(this, "인터넷 연결을 확인해주세요.", Toast.LENGTH_LONG).show();
+            return;
+        }
 
-        Call<Result> call = getCommentService().deleteComment(uId, commentId);
+        Call<Result> call = getCommentService().deleteComment(commentId);
 
         call.enqueue(new Callback<Result>() {
             @Override
             public void onResponse(Call<Result> call, Response<Result> response) {
                 if (response.isSuccessful()) {
-                    // deleteReportedComment(position, commentId);
                     snack("신고된 댓글을 삭제했습니다");
-                    reportedCommentAdapter.deleteComment(position);
+                    deleteReportedComment(position, commentId, false);
+
                 }
             }
 
@@ -530,7 +484,12 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
 
     // 댓글 삭제시 신고된 댓글도 삭제
     // TODO 아직 삭제가 안
-    private void deleteReportedComment(final int position, int commentId) {
+    private void deleteReportedComment(final int position, int commentId, final boolean refuseReport) {
+
+        if (!NetworkUtil.isConnected(this)) {
+            Toast.makeText(this, "인터넷 연결을 확인해주세요.", Toast.LENGTH_LONG).show();
+            return;
+        }
 
         Call<Result> call = getCommentService().deleteReportedComment(commentId);
 
@@ -538,7 +497,7 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
             @Override
             public void onResponse(Call<Result> call, Response<Result> response) {
                 if (response.isSuccessful()) {
-                    snack("신고를 취소했습니다");
+                    if (refuseReport) snack("신고를 취소했습니다");
                     reportedCommentAdapter.deleteComment(position);
                 }
             }
@@ -557,7 +516,7 @@ public class MyCommentActivity extends AppCompatActivity implements MyCommentExp
                 .setPositiveButton(getText(R.string.refuse_report), new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        deleteReportedComment(position, commentId);
+                        deleteReportedComment(position, commentId, true);
                     }
                 })
                 .setNegativeButton(getText(R.string.delete_comment), new DialogInterface.OnClickListener() {
